添加指定文件夹和工程文件 添加ARC编译参数

+++ Util/NSString+Extension.m
@@ -68,7 +68,7 @@
 - (NSString *)letters
 {
     NSMutableString *letterString = [NSMutableString string];
-    int len = [self length];
+    NSInteger len = [self length];
     for (int i = 0;i < len;i++)
     {
         NSString *oneChar = [[self substringFromIndex:i] substringToIndex:1];
